Use instance variables instead of options hash
There are only so many parts of the options hash that
gets passed in that we seem to care about, and since
we're working inside an object, it's a good idea to set
the objects we want to actually work with early on instead
of constantly (and possibly inconsistently) overriding or
defaulting. Tested and works find in Rails 4, Ruby 2.0.

This also prevents us from passing our data keys onto the
form input. I was seeing 'key_starts_with' as an attribute
on my inputs when I was overriding that option, and it
wasn't prefixed by 'data-'.